대나무숲 글 등록, 감정분석 구현

See merge request ai_track/class_07/ai_project/team05/mbti-agora!115

+++ back/src/controllers/storyCommentController.js
@@ -7,6 +7,9 @@ class StoryCommentController {
       const storyId = req.params.storyId;
       const writerId = req.currentUserId;
       const { comment } = req.body;
+      if (!comment) {
+        throw new Error('댓글을 입력해주세요');
+      }
       const obj = await axios.post('http://127.0.0.1:5000/predict', {
         text: comment,
       });
@@ -31,7 +34,18 @@ class StoryCommentController {
   static async updateStoryComment(req, res, next) {
     try {
       const commentId = req.params.commentId;
+      const loginUserId = req.currentUserId;
+      const isSameUser = await StoryCommentService.isSameUser(
+        loginUserId,
+        commentId,
+      );
+      if (!isSameUser) {
+        throw new Error('댓글 수정 권한이 없습니다.');
+      }
       const { comment } = req.body;
+      if (!comment) {
+        throw new Error('댓글을 입력해주세요');
+      }
       const obj = await axios.post('http://127.0.0.1:5000/predict', {
         text: comment,
       });
@@ -56,10 +70,41 @@ class StoryCommentController {
   static async deleteStoryComment(req, res, next) {
     try {
       const commentId = req.params.commentId;
+      const loginUserId = req.currentUserId;
+      const isSameUser = await StoryCommentService.isSameUser(
+        loginUserId,
+        commentId,
+      );
+      if (!isSameUser) {
+        throw new Error('댓글 삭제 권한이 없습니다.');
+      }
       const result = await StoryCommentService.deleteStoryComment({
         commentId,
       });
       return res.status(200).send(result);
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  static async readStoryComment(req, res, next) {
+    try {
+      const storyId = req.params.storyId;
+      const page = parseInt(req.query.page || 1);
+      const limit = 6;
+      const { comments, totalPage, count } =
+        await StoryCommentService.readComments(limit, page, storyId);
+      const populageResult = await StoryCommentService.populateStoryComment(
+        comments,
+        'writerId',
+      );
+      const result = {
+        currentPage: page,
+        totalPage: totalPage,
+        totalCommentsCount: count,
+        comments: populageResult,
+      };
+      return res.status(200).json(result);
     } catch (error) {
       next(error);
     }
